SPRC/tema03/mqttadapter/requirements.txt
influxdb==5.2.1
paho-mqtt==1.4.0
python-dateutil==2.8.1
